Merged PR 3682: Decoupled Splitter Hit Target & Visual Appearance
## Decoupled Splitter Hit Target & Visual Appearance

- Added new constant to splitter config `SPLITTER_HIT_TARGET`
  - Controls the size of the splitter handle. (`SPLITTER_HIT_TARGET = 8` means 4px of grab-able space on each side of the splitter's center)
- Splitter hit target is invisible and floated on top of the visual splitter

---

## Explanation for z-index changes

Initially, I didn't want to change the `z-index` of the split panes and the splitter to achieve this functionality, but due to the way the split panes work, it was the only way to get a centered hit target on top of the visual splitter without being cut off.

Currently, the way a splitter lays out its elements is that it creates two panes and a splitter div between them. Inside each pane is a floating, `position: absolute` canvas that takes up the entire space of the pane and contains the element that was supposed to be split. The splitter also floats a `position: absolute` div inside of it which acts as the invisible hit target.

**Example**

    <div
        <Splitter
            <div1
            <div2
        </Splitter>
    </div>

translates to:

    <div
        <div class="split-container"

            <!-- pane 1 -->
            <div class="split-pane
                <div class="floating-canvas"
                    <div1

            <!-- splitter -->
            <div class="splitter
                <div class="splitter-hit-target"

            <!-- pane 2 -->
            <div class="split-pane
                <div class="floating-canvas"
                    <div2
        </div>
    </div>

Due to the way stacking context works in HTML, `pane 2` and its contents are the last in the HTML markup so they are highest in the stacking context. This means that when the `splitter`'s floating hit target overlaps on `pane 2`'s side of the splitter, it becomes hidden under the contents of `pane 2` and therefore un-clickable. In order to solve this issue, I bumped up the `z-index` of the `splitter` to `1`, and each of the panes to `0` so that the splitter's hit target would float on top of both of them, and the stacking context would otherwise remain the same for the panes.

This also shouldn't cause any side effects with the flow of the page because the splitter component already wraps both panes and the splitter up into a containing `div.splitter-container` that has its own stacking context.