Merged PR 4415: Allow client-only extensions
NOTE: You must `lerna bootstrap` after pulling this change.

Extensions can now consist of a node process and a client build as separate projects. Both are optional. See the debug extension for an example.

I've updated the bf-extension.json schema to include separate configurations for node and client parts. Here's an example:

```json
{
  "name": "My Cool Extension",
  "node": {
    "main": "main/built/index.js",
    "debug": {
      "enabled": false,
      "websocket": {
        "port": 8081
      }
    }
  },
  "client": {
    "debug": {
      "enabled": false,
      "webpack": {
        "port": 3031
      }
    },
    "portals": [],
    "filetypes": [],
    "inspectors": [
      {
        "name": "Inspect Cool Things",
        "path": "client/public/index.html"
      }
    ]
  }
}
```

Notice the "debug" sections. I've added per-extension debug configuration to let the emulator know whether this extension is being debugged and how to load it if so.

##### node.debug section
The node debug section configures which port the emulator should listen on for connections from this extension.

##### client.debug section
The client debug section configures which port the webpack dev server is running on for this extension.