Merged PR 4765: Fixed bug where couldn't add more than one service in a session
Some of the service dialogs weren't populating the `id` field with a unique value, so it was defaulting to `undefined`. This meant that any time you would try to add more than one service, msbot would look through the services, find an existing service with `id: undefined`, and error out saying that a service with that id already exists.

Related work items: #22478