Merged PR 4917: Clicking on an endpoint should open to the right of the active editor
This handles opening the endpoint to the right of the active editor. There some stuff I don't like about the way I'm handling scrolling.  If anyone can show me how to figure out the width of an element instead of using a constant, that would be cleaner. And if there's a better way to find the index of the active tab, that would sure be nice.

Related work items: #22658