a11y: platform fixes for mac (#2381)
* a11y: 63767 - fix VoiceOver reading whole list content

In Electron/Chrome it reads the whole list content when the list root focused. Add a label to fix the issue.

* a11y: 64009 - improve left pane contrast

Fixes secondary text contrast and improves contrast ratio for primary text for the Left pane.
The bug mentioned some menu contrast issues, I believe the're not existent.
The bug also mentioned JSON contrast. It is already fixed.

* a11y: 64005 - make consistent full-screen handling between all platforms

Includes the following changes:
- Move the popup logic into the command
- Use the command for the native menu

* a11y: switch theme: announce on all platforms

Adds announcement for macOS as well. And makes message to start with visible content.

* Revert "a11y: 64005 - make consistent full-screen handling between all platforms"

This reverts commit 850cd46f32d718efe6cc5a5ec4db9e37f56f033d.

* Revert "a11y: switch theme: announce on all platforms"

This reverts commit 14aea00b857b63ce0eb60cdd35a13dfb7000b9f8.

* a11y: 63887 - make ngrok status accessible

Not able to reproduce the original issue, but found that the ngrok status is not being announced, so added the tunnel status and tunnel errors announcement.

* a11y: 63914 -  rework collapsable content announcement

Prevent lists from being announced when a new entry gets added.

Co-authored-by: Eugene Olonov <v-evolo@microsoft.com>
Co-authored-by: Tony Anziano <tonyanziano5@gmail.com>